Create MyQuery.java
Example console output (see testfile.txt):

Here is where your .txt file(s) should be located for this program to work properly:
C:\Users\[USER]\eclipse-workspace\MyQuery

Enter the filename that you want to Search values for. For example "MyQuery.txt"
testfile.txt
The file was found :) Here are the contents:
{skunks=[I don't like skunks], I=[I like food, I like video games, I like programming, I don't like skunks], video=[I like video games], like=[I like food, I like video games, I like programming, I don't like skunks], programming=[I like programming], don't=[I don't like skunks], food=[I like food], games=[I like video games]}
Enter a key to search for the value it is associated with.

like
[I like food, I like video games, I like programming, I don't like skunks]

Want to try again? If so, press return and then follow the prompt. Type "-1" to quit

Enter a key to search for the value it is associated with.

don't
[I don't like skunks]

Want to try again? If so, press return and then follow the prompt. Type "-1" to quit
-1